Allow configuration of acceptance test frameworks
This commit decouples testing frameworks (rspec, minitest) from acceptance test
frameworks (capybara, active support test case).

It does so by allowing one to determine which acceptance test framework
to use via a configuration option passed on to the gem.

As an example, one could set it to be `capybara` like this:

```
GovukAbTesting.configure do |config|
  config.acceptance_test_framework = :capybara
end
```

The gem will then create the relevant instance of the acceptance test
class and let it be used by either minitest or rspec.

In this change, I've also:
- Renamed RSpec helper, as this is no longer tied to a particular acceptance
  test framework;
- Use the acceptance test classes to fetch data: make sure we only fetch
  page-specific data from the acceptance test frameworks;
- Add documentation on new behaviour in the README.