Modify assertions to require A/B test name
This commit modifies the following assertions in both `Minitest` and
`RSpec` classes:

- `assert_response_is_cached_by_variant`
- `assert_response_not_modified_for_ab_test`

Both assertions now require a new parameter `ab_test_name`, which
indicates what A/B test we are running the assertions against.

This change is needed in case we are running more than 1 A/B test in a
given page.

This commit also introduces a `MetaTag` class which allows us to query
if a given meta tag belongs to a particular A/B test.