Update Ruby to version 3.2
Ruby version 2.7.6 is EOL on 1 April 2023.

In line with the advice on [Ruby version compatibility](https://docs.publishing.service.gov.uk/manual/publishing-a-ruby-gem.html#ruby-version-compatibility) and [testing multiple versions of Ruby in a gem](https://docs.publishing.service.gov.uk/manual/test-and-build-a-project-with-github-actions.html#a-ruby-gem), v2.7.6 has been removed and updated to v3.2.